requirements.txt
# AgencyGen - Meta-Agent System Generator
# ========================================
# Uses Google's Agent Development Kit (ADK)
# Documentation: https://google.github.io/adk-docs/

# Google ADK - The core framework
google-adk>=1.0.0

# Note: google-adk includes google-genai as a dependency

# To run:
# 1. pip install -r requirements.txt
# 2. export GOOGLE_API_KEY=your_key_here
# 3. python example.py
